Extract fetching main graph and remapping main graph data
